python/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asn1crypto==1.5.1
bip-utils==2.9.3
bip32==4.0
bitcoinlib==0.7.5
cbor2==5.6.5
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
coincurve==20.0.0
crcmod==1.7
cryptography==46.0.1
ecdsa==0.19.1
ed25519-blake2b==1.4.1
exceptiongroup==1.3.0
fastecdsa==2.3.2
greenlet==3.1.1
idna==3.10
iniconfig==2.1.0
mnemonic==0.21
numpy==1.24.4
packaging==25.0
pillow==10.4.0
pluggy==1.5.0
py-sr25519-bindings==0.2.2
pycparser==2.23
pycryptodome==3.23.0
pynacl==1.6.0
pytest==8.3.5
python-bitcoinrpc==1.0
python-bitcointx==1.1.5
python-dotenv==1.0.1
pyzbar==0.1.9
requests==2.32.4
six==1.17.0
sqlalchemy==2.0.43
tomli==2.2.1
typing-extensions==4.13.2
urllib3==2.2.3
